MarketX-backend/ecommerce_api_project/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
psycopg2-binary==2.9.1
djangorestframework_simplejwt==4.8.0
django-filter==2.4.0
